+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Properly share the requested IntelliJ Platform between internal Gradle plugin components
 - Refactor `CollectorTransformer` to exclude tests-related jars using `filterNot` as `minus` doesn't remove duplicates
 - Refactor `JarSearchableOptionsTask` to improve the `searchableOption` outgoing artifact registration JetBrains/intellij-platform-gradle-plugin#1968
+- Do not add default dependencies related to the code instrumentation if disabled JetBrains/intellij-platform-gradle-plugin#1983
 
 ## [2.6.0] - 2025-05-14